refactor : 프로젝트 패키지 구조 개선 및 경로 수정
