+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Background/BackgroundRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.techeerpicture.TecheerPicture.Image.Image;
+import com.techeerpicture.TecheerPicture.Image.entity.Image;
 import java.util.Optional;
 import java.util.List;
