+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Background/BackgroundService.java
@@ -9,8 +9,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.techeerpicture.TecheerPicture.Background.PixelcutService;
 import com.techeerpicture.TecheerPicture.Background.Background;
 import com.techeerpicture.TecheerPicture.Background.BackgroundRepository;
-import com.techeerpicture.TecheerPicture.Image.Image;
-import com.techeerpicture.TecheerPicture.Image.ImageRepository;
+import com.techeerpicture.TecheerPicture.Image.entity.Image;
+import com.techeerpicture.TecheerPicture.Image.repository.ImageRepository;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import java.net.MalformedURLException;
