+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Banner/entity/Banner.java
@@ -1,6 +1,7 @@
 package com.techeerpicture.TecheerPicture.Banner.entity;
 
-import com.techeerpicture.TecheerPicture.Image.Image;
+
+import com.techeerpicture.TecheerPicture.Image.entity.Image;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import jakarta.persistence.*;
