+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/ImageToVideo/ImageToVideoService.java
@@ -5,8 +5,8 @@ import ai.fal.client.queue.*;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.google.gson.JsonObject;
-import com.techeerpicture.TecheerPicture.Image.Image;
-import com.techeerpicture.TecheerPicture.Image.ImageRepository;
+import com.techeerpicture.TecheerPicture.Image.entity.Image;
+import com.techeerpicture.TecheerPicture.Image.repository.ImageRepository;
 import com.techeerpicture.TecheerPicture.Background.TranslationService;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
