+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/RemoveBackground/service/RemoveBackgroundService.java
@@ -4,8 +4,9 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.techeerpicture.TecheerPicture.Background.Background;
 import com.techeerpicture.TecheerPicture.Background.BackgroundRepository;
-import com.techeerpicture.TecheerPicture.Image.Image;
-import com.techeerpicture.TecheerPicture.Image.ImageRepository;
+import com.techeerpicture.TecheerPicture.Image.entity.Image;
+import com.techeerpicture.TecheerPicture.Image.repository.ImageRepository;
+
 import com.techeerpicture.TecheerPicture.RemoveBackground.dto.RemoveBackgroundResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
